Add files via upload
Variant of standard_oddball with temporal jitter between stimuli: This Bonsai workflow is a modified version of the standard_oddball visual oddball paradigm, incorporating a temporal jitter between stimulus presentations. The inter stimulus intervals (delay parameter) are uniformly randomly sampled from 0.343s, 1s, 1.5 s, and 2 s to introduce variability in the timing of the stimuli. The jitter values are specified together with the orientations in the Grating Specification section of the workflow.